test(cli): extract dispatch_surface_gap tests to sidecar
The merge surfaced a CI module-size failure: the branch had added a
~170-line inline #[cfg(test)] mod tests to src/cli/dispatch_surface_gap.rs,
pushing it to 595 counted lines (>500 limit) and violating the repo's
sidecar-test convention. Move the four argument-mapper tests into
src/cli/dispatch_surface_gap_tests.rs behind the standard
#[path = "..."] mod tests; hook. No behavior change.
